+++ actix-files/src/lib.rs
@@ -83,7 +83,7 @@ mod tests {
 
     use super::*;
 
-    #[actix_rt::test]
+    #[actix_web::test]
     async fn test_file_extension_to_mime() {
         let m = file_extension_to_mime("");
         assert_eq!(m, mime::APPLICATION_OCTET_STREAM);