+++ actix-web-codegen/tests/trybuild.rs
@@ -1,4 +1,4 @@
-#[rustversion::stable(1.51)] // MSRV
+#[rustversion::stable(1.52)] // MSRV
 #[test]
 fn compile_macros() {
     let t = trybuild::TestCases::new();
@@ -13,4 +13,6 @@ fn compile_macros() {
     t.compile_fail("tests/trybuild/route-malformed-path-fail.rs");
 
     t.pass("tests/trybuild/docstring-ok.rs");
+
+    t.pass("tests/trybuild/test-runtime.rs");
 }