+++ awc/src/client/connection.rs
@@ -12,10 +12,9 @@ use bytes::Bytes;
 use futures_core::future::LocalBoxFuture;
 use h2::client::SendRequest;
 
-use crate::h1::ClientCodec;
-use crate::message::{RequestHeadType, ResponseHead};
-use crate::payload::Payload;
-use crate::{body::MessageBody, Error};
+use actix_http::{
+    body::MessageBody, h1::ClientCodec, Error, Payload, RequestHeadType, ResponseHead,
+};
 
 use super::error::SendRequestError;
 use super::pool::Acquired;
@@ -219,11 +218,7 @@ impl<Io: ConnectionIo> ConnectionType<Io> {
         }
     }
 
-    pub(super) fn from_h1(
-        io: Io,
-        created: time::Instant,
-        acquired: Acquired<Io>,
-    ) -> Self {
+    pub(super) fn from_h1(io: Io, created: time::Instant, acquired: Acquired<Io>) -> Self {
         Self::H1(H1Connection {
             io: Some(io),
             created,
@@ -271,9 +266,7 @@ where
                 Connection::Tls(ConnectionType::H2(conn)) => {
                     h2proto::send_request(conn, head.into(), body).await
                 }
-                _ => unreachable!(
-                    "Plain Tcp connection can be used only in Http1 protocol"
-                ),
+                _ => unreachable!("Plain Tcp connection can be used only in Http1 protocol"),
             }
         })
     }
@@ -301,9 +294,7 @@ where
                     Err(SendRequestError::TunnelNotSupported)
                 }
                 Connection::Tcp(ConnectionType::H2(_)) => {
-                    unreachable!(
-                        "Plain Tcp connection can be used only in Http1 protocol"
-                    )
+                    unreachable!("Plain Tcp connection can be used only in Http1 protocol")
                 }
             }
         })
@@ -321,12 +312,8 @@ where
         buf: &mut ReadBuf<'_>,
     ) -> Poll<io::Result<()>> {
         match self.get_mut() {
-            Connection::Tcp(ConnectionType::H1(conn)) => {
-                Pin::new(conn).poll_read(cx, buf)
-            }
-            Connection::Tls(ConnectionType::H1(conn)) => {
-                Pin::new(conn).poll_read(cx, buf)
-            }
+            Connection::Tcp(ConnectionType::H1(conn)) => Pin::new(conn).poll_read(cx, buf),
+            Connection::Tls(ConnectionType::H1(conn)) => Pin::new(conn).poll_read(cx, buf),
             _ => unreachable!("H2Connection can not impl AsyncRead trait"),
         }
     }
@@ -345,12 +332,8 @@ where
         buf: &[u8],
     ) -> Poll<io::Result<usize>> {
         match self.get_mut() {
-            Connection::Tcp(ConnectionType::H1(conn)) => {
-                Pin::new(conn).poll_write(cx, buf)
-            }
-            Connection::Tls(ConnectionType::H1(conn)) => {
-                Pin::new(conn).poll_write(cx, buf)
-            }
+            Connection::Tcp(ConnectionType::H1(conn)) => Pin::new(conn).poll_write(cx, buf),
+            Connection::Tls(ConnectionType::H1(conn)) => Pin::new(conn).poll_write(cx, buf),
             _ => unreachable!(H2_UNREACHABLE_WRITE),
         }
     }
@@ -363,17 +346,10 @@ where
         }
     }
 
-    fn poll_shutdown(
-        self: Pin<&mut Self>,
-        cx: &mut Context<'_>,
-    ) -> Poll<io::Result<()>> {
+    fn poll_shutdown(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
         match self.get_mut() {
-            Connection::Tcp(ConnectionType::H1(conn)) => {
-                Pin::new(conn).poll_shutdown(cx)
-            }
-            Connection::Tls(ConnectionType::H1(conn)) => {
-                Pin::new(conn).poll_shutdown(cx)
-            }
+            Connection::Tcp(ConnectionType::H1(conn)) => Pin::new(conn).poll_shutdown(cx),
+            Connection::Tls(ConnectionType::H1(conn)) => Pin::new(conn).poll_shutdown(cx),
             _ => unreachable!(H2_UNREACHABLE_WRITE),
         }
     }