+++ actix-files/CHANGES.md
@@ -3,6 +3,10 @@
 ## Unreleased - 2021-xx-xx
 
 
+## 0.6.0-beta.8 - 2021-10-20
+* Minimum supported Rust version (MSRV) is now 1.52.
+
+
 ## 0.6.0-beta.7 - 2021-09-09
 * Minimum supported Rust version (MSRV) is now 1.51.