+++ actix-web-codegen/CHANGES.md
@@ -3,6 +3,15 @@
 ## Unreleased - 2021-xx-xx
 
 
+## 0.5.0-beta.5 - 2021-10-20
+* Improve error recovery potential when macro input is invalid. [#2410]
+* Add `#[actix_web::test]` macro for setting up tests with a runtime. [#2409]
+* Minimum supported Rust version (MSRV) is now 1.52.
+
+[#2410]: https://github.com/actix/actix-web/pull/2410
+[#2409]: https://github.com/actix/actix-web/pull/2409
+
+
 ## 0.5.0-beta.4 - 2021-09-09
 * In routing macros, paths are now validated at compile time. [#2350]
 * Minimum supported Rust version (MSRV) is now 1.51.